Adding app.py and fetch.py
app.py is a Streamlit application that serves as an Air Quality Dashboard. It connects to a MongoDB database hosted on MongoDB Atlas to fetch and display real-time air quality data for major cities across the USA. The dashboard provides interactive visualizations using Plotly Express, allowing users to explore the latest air quality metrics and historical trends for each city.

fetch.py
fetch.py is a Python script responsible for fetching air quality data from the AirVisual API. It retrieves current pollution levels, weather conditions, and other relevant metrics for a predefined list of major cities in the USA. The fetched data is then stored in a MongoDB database hosted on MongoDB Atlas, ensuring that the dashboard (app.py) has access to the most up-to-date information for visualizations and analysis.